refactor: Tree traversals now non-recursive
