+++ bigframes/core/nodes.py
@@ -20,7 +20,7 @@ import datetime
 import functools
 import itertools
 import typing
-from typing import Callable, cast, Iterable, Mapping, Optional, Sequence, Tuple, TypeVar
+from typing import Callable, cast, Iterable, Mapping, Optional, Sequence, Tuple
 
 import google.cloud.bigquery as bq
 
@@ -28,6 +28,7 @@ from bigframes.core import identifiers
 import bigframes.core.expression as ex
 import bigframes.core.guid
 from bigframes.core.ordering import OrderingExpression
+from bigframes.core.plan import BigFrameNode, COLUMN_SET, Field
 import bigframes.core.schema as schemata
 import bigframes.core.slices as slices
 import bigframes.core.window_spec as window
@@ -40,246 +41,6 @@ if typing.TYPE_CHECKING:
 
 # A fixed number of variable to assume for overhead on some operations
 OVERHEAD_VARIABLES = 5
-
-COLUMN_SET = frozenset[identifiers.ColumnId]
-
-Self = TypeVar("Self")
-
-
-@dataclasses.dataclass(frozen=True)
-class Field:
-    id: identifiers.ColumnId
-    dtype: bigframes.dtypes.Dtype
-
-
-@dataclasses.dataclass(eq=False, frozen=True)
-class BigFrameNode(abc.ABC):
-    """
-    Immutable node for representing 2D typed array as a tree of operators.
-
-    All subclasses must be hashable so as to be usable as caching key.
-    """
-
-    @property
-    def deterministic(self) -> bool:
-        """Whether this node will evaluates deterministically."""
-        return True
-
-    @property
-    def row_preserving(self) -> bool:
-        """Whether this node preserves input rows."""
-        return True
-
-    @property
-    def non_local(self) -> bool:
-        """
-        Whether this node combines information across multiple rows instead of processing rows independently.
-        Used as an approximation for whether the expression may require shuffling to execute (and therefore be expensive).
-        """
-        return False
-
-    @property
-    def child_nodes(self) -> typing.Sequence[BigFrameNode]:
-        """Direct children of this node"""
-        return tuple([])
-
-    @property
-    @abc.abstractmethod
-    def row_count(self) -> typing.Optional[int]:
-        return None
-
-    @abc.abstractmethod
-    def remap_vars(
-        self: Self, mappings: Mapping[identifiers.ColumnId, identifiers.ColumnId]
-    ) -> Self:
-        """Remap defined (in this node only) variables."""
-        ...
-
-    @abc.abstractmethod
-    def remap_refs(
-        self: Self, mappings: Mapping[identifiers.ColumnId, identifiers.ColumnId]
-    ) -> Self:
-        """Remap variable references"""
-        ...
-
-    @property
-    @abc.abstractmethod
-    def node_defined_ids(self) -> Tuple[identifiers.ColumnId, ...]:
-        """The variables defined in this node (as opposed to by child nodes)."""
-        ...
-
-    @property
-    def referenced_ids(self) -> COLUMN_SET:
-        return frozenset()
-
-    @functools.cached_property
-    def session(self):
-        sessions = []
-        for child in self.child_nodes:
-            if child.session is not None:
-                sessions.append(child.session)
-        unique_sessions = len(set(sessions))
-        if unique_sessions > 1:
-            raise ValueError("Cannot use combine sources from multiple sessions.")
-        elif unique_sessions == 1:
-            return sessions[0]
-        return None
-
-    def _validate(self):
-        """Validate the local data in the node."""
-        return
-
-    @functools.cache
-    def validate_tree(self) -> bool:
-        for child in self.child_nodes:
-            child.validate_tree()
-        self._validate()
-        field_list = list(self.fields)
-        if len(set(field_list)) != len(field_list):
-            raise ValueError(f"Non unique field ids {list(self.fields)}")
-        return True
-
-    def _as_tuple(self) -> Tuple:
-        """Get all fields as tuple."""
-        return tuple(getattr(self, field.name) for field in dataclasses.fields(self))
-
-    def __hash__(self) -> int:
-        # Custom hash that uses cache to avoid costly recomputation
-        return self._cached_hash
-
-    def __eq__(self, other) -> bool:
-        # Custom eq that tries to short-circuit full structural comparison
-        if not isinstance(other, self.__class__):
-            return False
-        if self is other:
-            return True
-        if hash(self) != hash(other):
-            return False
-        return self._as_tuple() == other._as_tuple()
-
-    # BigFrameNode trees can be very deep so its important avoid recalculating the hash from scratch
-    # Each subclass of BigFrameNode should use this property to implement __hash__
-    # The default dataclass-generated __hash__ method is not cached
-    @functools.cached_property
-    def _cached_hash(self):
-        return hash(self._as_tuple())
-
-    @property
-    def roots(self) -> typing.Set[BigFrameNode]:
-        roots = itertools.chain.from_iterable(
-            map(lambda child: child.roots, self.child_nodes)
-        )
-        return set(roots)
-
-    # TODO: Store some local data lazily for select, aggregate nodes.
-    @property
-    @abc.abstractmethod
-    def fields(self) -> Iterable[Field]:
-        ...
-
-    @property
-    def ids(self) -> Iterable[identifiers.ColumnId]:
-        """All output ids from the node."""
-        return (field.id for field in self.fields)
-
-    @property
-    @abc.abstractmethod
-    def variables_introduced(self) -> int:
-        """
-        Defines number of values created by the current node. Helps represent the "width" of a query
-        """
-        ...
-
-    @property
-    def relation_ops_created(self) -> int:
-        """
-        Defines the number of relational ops generated by the current node. Used to estimate query planning complexity.
-        """
-        return 1
-
-    @property
-    def joins(self) -> bool:
-        """
-        Defines whether the node joins data.
-        """
-        return False
-
-    @property
-    @abc.abstractmethod
-    def order_ambiguous(self) -> bool:
-        """
-        Whether row ordering is potentially ambiguous. For example, ReadTable (without a primary key) could be ordered in different ways.
-        """
-        ...
-
-    @property
-    @abc.abstractmethod
-    def explicitly_ordered(self) -> bool:
-        """
-        Whether row ordering is potentially ambiguous. For example, ReadTable (without a primary key) could be ordered in different ways.
-        """
-        ...
-
-    @functools.cached_property
-    def height(self) -> int:
-        if len(self.child_nodes) == 0:
-            return 0
-        return max(child.height for child in self.child_nodes) + 1
-
-    @functools.cached_property
-    def total_variables(self) -> int:
-        return self.variables_introduced + sum(
-            map(lambda x: x.total_variables, self.child_nodes)
-        )
-
-    @functools.cached_property
-    def total_relational_ops(self) -> int:
-        return self.relation_ops_created + sum(
-            map(lambda x: x.total_relational_ops, self.child_nodes)
-        )
-
-    @functools.cached_property
-    def total_joins(self) -> int:
-        return int(self.joins) + sum(map(lambda x: x.total_joins, self.child_nodes))
-
-    @functools.cached_property
-    def schema(self) -> schemata.ArraySchema:
-        # TODO: Make schema just a view on fields
-        return schemata.ArraySchema(
-            tuple(schemata.SchemaItem(i.id.name, i.dtype) for i in self.fields)
-        )
-
-    @property
-    def planning_complexity(self) -> int:
-        """
-        Empirical heuristic measure of planning complexity.
-
-        Used to determine when to decompose overly complex computations. May require tuning.
-        """
-        return self.total_variables * self.total_relational_ops * (1 + self.total_joins)
-
-    @abc.abstractmethod
-    def transform_children(
-        self: Self, t: Callable[[BigFrameNode], BigFrameNode]
-    ) -> Self:
-        """Apply a function to each child node."""
-        ...
-
-    @property
-    def defines_namespace(self) -> bool:
-        """
-        If true, this node establishes a new column id namespace.
-
-        If false, this node consumes and produces ids in the namespace
-        """
-        return False
-
-    def get_type(self, id: identifiers.ColumnId) -> bigframes.dtypes.Dtype:
-        return self._dtype_lookup[id]
-
-    @functools.cached_property
-    def _dtype_lookup(self):
-        return {field.id: field.dtype for field in self.fields}
 
 
 class AdditiveNode:
@@ -1745,50 +1506,24 @@ class ExplodeNode(UnaryNode):
 def top_down(
     root: BigFrameNode,
     transform: Callable[[BigFrameNode], BigFrameNode],
-    *,
-    memoize=False,
-    validate=False,
 ) -> BigFrameNode:
     """
     Perform a top-down transformation of the BigFrameNode tree.
-
-    If memoize=True, recursive calls are memoized within the scope of the traversal only.
     """
-
-    def top_down_internal(root: BigFrameNode) -> BigFrameNode:
-        return transform(root).transform_children(top_down_internal)
-
-    if memoize:
-        # MUST reassign to the same name or caching won't work recursively
-        top_down_internal = functools.cache(top_down_internal)
-
-    result = top_down_internal(root)
-    if validate:
-        result.validate_tree()
-    return result
+    return root.top_down(transform)
 
 
 def bottom_up(
     root: BigFrameNode,
     transform: Callable[[BigFrameNode], BigFrameNode],
-    *,
-    memoize=False,
-    validate=False,
 ) -> BigFrameNode:
     """
     Perform a bottom-up transformation of the BigFrameNode tree.
 
-    If memoize=True, recursive calls are memoized within the scope of the traversal only.
+    The `transform` function is applied to each node *after* its children
+    have been transformed.  This allows for transformations that depend
+    on the results of transforming subtrees.
+
+    Returns the transformed root node.
     """
-
-    def bottom_up_internal(root: BigFrameNode) -> BigFrameNode:
-        return transform(root.transform_children(bottom_up_internal))
-
-    if memoize:
-        # MUST reassign to the same name or caching won't work recursively
-        bottom_up_internal = functools.cache(bottom_up_internal)
-
-    result = bottom_up_internal(root)
-    if validate:
-        result.validate_tree()
-    return result
+    return root.bottom_up(transform)
